Atualização fbanco de dados, controller e model

+++ app/Controllers/Http/FaturaController.js
@@ -5,7 +5,13 @@
 /** @typedef {import('@adonisjs/framework/src/View')} View */
 const Usuario = use('App/Models/User');
 const Fatura = use('App/Models/Fatura');
-const {format, subDays} = use('date-fns')
+const {format, subDays} = use('date-fns');
+const Asaas = use('App/Models/Asaas/Asaas');
+
+const Env = use('Env')
+const SANDBOX_ASAAS = Env.get('SANDBOX_ASAAS')
+const ACCESS_TOKEN_ASAAS_SANDBOX = Env.get('ACCESS_TOKEN_ASAAS_SANDBOX')
+const ACCESS_TOKEN_ASAAS_PRODUCION = Env.get('ACCESS_TOKEN_ASAAS_PRODUCION');
 /**
  * Resourceful controller for interacting with faturas
  */
@@ -56,11 +62,33 @@ class FaturaController {
   }
 
   async store ({ request, response }) {
+
     const data = request.only([
-      'user_id', 'servico_id', 'servico_clientes_id', 'vencimento', 'valor', 'obs', 'status',
+      'user_id', 'servico_id', 'servico_cliente_id', 'id_integracao', 'vencimento', 'valor', 'link_fartura', 'link_fartura_aux', 'obs', 'status', 'cpfCnpj'
     ]);
 
-    // console.log("data cadastro", data)
+    console.log('data store fatura', data)
+
+    const dataCliente = {
+      'cpfCnpj' : data.cpfCnpj
+    }
+
+    // //buscar cliente no asaas
+    // let access_token = ""
+    // access_token = '8d695fbb569a6e3a5e7269fee1a437a2a78db7d10abfce52c8b1dd5da6c99966'
+    // var retorno = await Asaas.consultarCliente(access_token, dataCliente);
+    // console.log("retorno verficar cliente ", retorno)
+
+    // const qtoClienteAsaas = retorno.data.length;
+    // console.log("qtoClienteAsaas", qtoClienteAsaas)
+    
+    // async function criarCobrancasAsaas(data, elementServico, mesAnoRef){
+    //   let access_token = ""
+    //   access_token = '8d695fbb569a6e3a5e7269fee1a437a2a78db7d10abfce52c8b1dd5da6c99966'
+    //   var retorno = await Asaas.criarCobranca(access_token, data, elementServico);
+    //   // console.log('Criando cobrança asaas retorno', retorno)
+    //   salavarCobranca(retorno, elementServico, mesAnoRef)
+    // }
 
     //buscar o dia atual
     const dataAtual = format(new Date(data.vencimento), "yyyy-MM-dd HH:mm:ss")
@@ -74,11 +102,14 @@ class FaturaController {
     const fatura = await Fatura.create({
       user_id: data.user_id, // id do cliente
       servico_id: data.servico_id,
-      servico_clientes_id: data.servico_clientes_id,
+      servico_cliente_id: data.servico_cliente_id,
+      id_integracao: data.id_integracao,
       mes_referencia: mesRef,
       ano_referencia: anoRef,
       vencimento: data.vencimento,
       valor: valor,
+      link_fartura: data.link_fartura,
+      link_fartura_aux: data.link_fartura_aux,
       obs: data.obs,
       status: data.status,
     });
@@ -105,7 +136,7 @@ class FaturaController {
       const fatura = await Fatura.find(params.id);
 
       let data = request.only([
-        'data_pagamento', 'status', 'obs'
+        'data_pagamento', 'valor_liquido', 'metodo_pagamento', 'status', 'obs',
       ]);
       console.log("Update fatura", data)
 
@@ -114,6 +145,8 @@ class FaturaController {
 
       fatura.merge({
         data_pagamento: data['data_pagamento'],
+        valor_liquido: data['valor_liquido'],
+        metodo_pagamento: data['metodo_pagamento'],
         status: data['status'],
         obs: data['obs']
       })
@@ -131,7 +164,27 @@ class FaturaController {
     const roles = await usuario.getRoles()
     if(roles[0] === "admin"){
       const fatura = await Fatura.findByOrFail('id', params.id);
-      fatura.delete();
+      const faturaJson = fatura.toJSON();
+      if(faturaJson.id_integracao){
+        console.log('deletando fatura, com integração', faturaJson.id_integracao)
+        deletandoAsaas(faturaJson.id_integracao)
+      }else{
+        console.log('deletando fatura sem integração', fatura.toJSON())
+        fatura.delete();
+      }
+      
+      async function deletandoAsaas(id){
+        let access_token = ""
+        access_token = SANDBOX_ASAAS ? ACCESS_TOKEN_ASAAS_SANDBOX : ACCESS_TOKEN_ASAAS_PRODUCION
+        // var id = 'pay_6152857367477791'
+        var retorno = await Asaas.deletarCobranca(access_token, id);
+        console.log('returno asaas 2', retorno)
+        if(retorno.deleted){
+          return fatura.delete();
+        }else{
+          return 204;
+        }
+      }
     }
   }
 }
